refactor(kiloclaw): extract pure helpers from kiloclaw-instance DO
Extract module-level pure functions and types into focused files:
- regions.ts: parseRegions, shuffleRegions, deprioritizeRegion
- machine-recovery.ts: selectRecoveryCandidate, volumeIdFromMachine, recovery constants
- machine-config.ts: buildMachineConfig, guestFromSize, volumeNameFromSandboxId, metadata keys
- gateway-controller-types.ts: GatewayProcessStatus, GatewayControllerError, Zod schemas

Main DO file: 2926 → 2714 lines (-212). No behavioral changes.
