+++ kiloclaw/src/durable-objects/gateway-controller-types.ts
@@ -1,0 +1,53 @@
+import { z, type ZodType } from 'zod';
+
+export type GatewayProcessStatus = {
+  state: 'stopped' | 'starting' | 'running' | 'stopping' | 'crashed' | 'shutting_down';
+  pid: number | null;
+  uptime: number;
+  restarts: number;
+  lastExit: {
+    code: number | null;
+    signal: NodeJS.Signals | null;
+    at: string;
+  } | null;
+};
+
+export const GatewayProcessStatusSchema: ZodType<GatewayProcessStatus> = z.object({
+  state: z.enum(['stopped', 'starting', 'running', 'stopping', 'crashed', 'shutting_down']),
+  pid: z.number().int().nullable(),
+  uptime: z.number(),
+  restarts: z.number().int(),
+  lastExit: z
+    .object({
+      code: z.number().int().nullable(),
+      signal: z
+        .custom<NodeJS.Signals>((value): value is NodeJS.Signals => typeof value === 'string')
+        .nullable(),
+      at: z.string(),
+    })
+    .nullable(),
+});
+
+export const GatewayCommandResponseSchema = z.object({
+  ok: z.boolean(),
+});
+
+export const ConfigRestoreResponseSchema = z.object({
+  ok: z.boolean(),
+  signaled: z.boolean(),
+});
+
+export const ControllerVersionResponseSchema = z.object({
+  version: z.string(),
+  commit: z.string(),
+});
+
+export class GatewayControllerError extends Error {
+  readonly status: number;
+
+  constructor(status: number, message: string) {
+    super(message);
+    this.name = 'GatewayControllerError';
+    this.status = status;
+  }
+}
